+++ src/BDSOutputROOTEventTrajectory.cc
@@ -31,6 +31,8 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "BDSPhysicalVolumeInfoRegistry.hh"
 #include "BDSPhysicalVolumeInfo.hh"
 #include "BDSTrajectory.hh"
+
+#include <cmath>
 #endif
 
 ClassImp(BDSOutputROOTEventTrajectory)
@@ -124,10 +126,10 @@ void BDSOutputROOTEventTrajectory::Fill(const BDSTrajectoriesToStore* trajectori
 	{continue;}
 
       partID.push_back((int &&) traj->GetPDGEncoding());
-      trackID.push_back((unsigned int &&) traj->GetTrackID());
-      parentID.push_back((unsigned int &&) traj->GetParentID());
-      parentIndex.push_back((int &&) traj->GetParentIndex());
-      parentStepIndex.push_back((int &&) traj->GetParentStepIndex());
+      trackID.push_back((unsigned int) std::abs(traj->GetTrackID()));
+      parentID.push_back((unsigned int) std::abs(traj->GetParentID()));
+      parentIndex.push_back((unsigned int) std::abs(traj->GetParentIndex()));
+      parentStepIndex.push_back((unsigned int) std::abs(traj->GetParentStepIndex()));
 
       // now we convert the geant4 type based BDSTrajectory information into
       // basic C++ and ROOT types for the output
@@ -490,13 +492,13 @@ BDSOutputROOTEventTrajectoryPoint BDSOutputROOTEventTrajectory::primaryProcessPo
 
 std::vector<BDSOutputROOTEventTrajectoryPoint> BDSOutputROOTEventTrajectory::processHistory(int trackid)
 {
-  int ti = trackID_trackIndex.at(trackid);  // get track index
+  unsigned int ti = (unsigned int) std::abs(trackID_trackIndex.at(trackid));  // get track index
 
   std::vector<BDSOutputROOTEventTrajectoryPoint> tpv;      // trajectory point vector
   while (ti != 0)
     {
-      int pi  = parentIndex.at(ti);
-      int psi = parentStepIndex.at(ti);
+      unsigned int pi  = parentIndex.at(ti);
+      unsigned int psi = parentStepIndex.at(ti);
       
       BDSOutputROOTEventTrajectoryPoint p(partID[pi],
 					  trackID[pi],